changed things to jsx and fixed the profile error
